bare minimum implementation of platform_sf2000, sf2000_gfx and sf2000_input drivers. retroarch loads and can navigate the menus.
sf2000_input pretends to be RETRO_DEVICE_KEYBOARD instead of RETRO_DEVICE_JOYPAD as it should.
need to fix that, or maybe just move its functionality to sf2000_joypad

sf2000_gfx needs alot of work.
only the a very simple blit is supported now, which seems to be enough to display the menu.